+++ src/main/kotlin/com/bugsnag/android/gradle/BugsnagUploadJsSourceMapTask.kt
@@ -1,17 +1,28 @@
 package com.bugsnag.android.gradle
 
+import com.bugsnag.android.gradle.internal.GradleVersions
+import com.bugsnag.android.gradle.internal.property
+import com.bugsnag.android.gradle.internal.register
+import com.bugsnag.android.gradle.internal.versionNumber
 import org.gradle.api.DefaultTask
+import org.gradle.api.Project
 import org.gradle.api.Task
+import org.gradle.api.file.ConfigurableFileCollection
+import org.gradle.api.file.ProjectLayout
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.model.ObjectFactory
+import org.gradle.api.provider.Property
+import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.InputFile
+import org.gradle.api.tasks.InputFiles
 import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.PathSensitive
 import org.gradle.api.tasks.PathSensitivity
 import org.gradle.api.tasks.TaskAction
+import org.gradle.api.tasks.TaskProvider
 import javax.inject.Inject
 
-open class BugsnagUploadJsSourceMapTask @Inject constructor(
+sealed class BugsnagUploadJsSourceMapTask @Inject constructor(
     objects: ObjectFactory
 ) : DefaultTask(), AndroidManifestInfoReceiver {
 
@@ -25,20 +36,85 @@ open class BugsnagUploadJsSourceMapTask @Inject constructor(
     override val manifestInfoFile: RegularFileProperty = objects.fileProperty()
 
     @get:InputFile
-    val bundleJsFile: RegularFileProperty = objects.fileProperty()
+    val bundleJsFileProvider: RegularFileProperty = objects.fileProperty()
 
     @get:InputFile
-    val sourceMapFile: RegularFileProperty = objects.fileProperty()
+    val sourceMapFileProvider: RegularFileProperty = objects.fileProperty()
 
     @get:OutputFile
     val requestOutputFile: RegularFileProperty = objects.fileProperty()
+
+    @get:InputFiles
+    abstract val projectRootFileProvider: ConfigurableFileCollection
+
+    @get:Input
+    val overwrite: Property<Boolean> = objects.property()
+
+    @get:Input
+    val endpoint: Property<String> = objects.property()
+
+    @get:Input
+    val devEnabled: Property<String> = objects.property()
+
+    @get:Input
+    val failOnUploadError: Property<Boolean> = objects.property()
 
     @TaskAction
     fun uploadJsSourceMap() {
         // Construct a basic request
         val manifestInfo = parseManifestInfo()
-        val cliResult = "success"
-        project.logger.lifecycle("Uploading sourcemap: ${bundleJsFile.get()}, JS bundle: ${bundleJsFile.get()}")
+
+        val builder = ProcessBuilder(
+            "echo", // TODO this is a placeholder until bugsnag-source-maps is ready
+
+            "--api-key",
+            manifestInfo.apiKey,
+
+            "--app-version",
+            manifestInfo.versionName,
+
+            "--app-version-code",
+            manifestInfo.versionCode,
+
+            "--dev",
+            devEnabled.get(),
+
+            "--platform",
+            "android",
+
+            "--source-map",
+            sourceMapFileProvider.asFile.get().absolutePath,
+
+            "--bundle",
+            bundleJsFileProvider.asFile.get().absolutePath,
+
+            "--overwrite",
+            overwrite.get().toString(),
+
+            "--endpoint",
+            endpoint.get(),
+
+            "--project-root",
+            projectRootFileProvider.singleFile.absolutePath
+        )
+        builder.redirectError(ProcessBuilder.Redirect.INHERIT)
+
+        project.logger.lifecycle("Bugsnag: uploading react native sourcemap: ${builder.command()}")
+
+        val process = builder.start()
+        val exitCode = process.waitFor()
+
+        if (exitCode != 0 && failOnUploadError.get()) {
+            throw IllegalStateException(
+                "Bugsnag: source map upload failed, $exitCode." +
+                    "Please ensure that bugsnag-source-maps is installed and source maps are enabled."
+            )
+        }
+
+        val cliResult = when (exitCode) {
+            0 -> "success"
+            else -> "failure"
+        }
         requestOutputFile.asFile.get().writeText(cliResult)
     }
 
@@ -56,5 +132,43 @@ open class BugsnagUploadJsSourceMapTask @Inject constructor(
             }
             return null
         }
+
+        /**
+         * Registers the appropriate subtype to this [project] with the given [name] and
+         * [configurationAction]
+         */
+        internal fun register(
+            project: Project,
+            name: String,
+            configurationAction: BugsnagUploadJsSourceMapTask.() -> Unit
+        ): TaskProvider<out BugsnagUploadJsSourceMapTask> {
+            return when {
+                project.gradle.versionNumber() >= GradleVersions.VERSION_5_3 -> {
+                    project.tasks.register<BugsnagUploadJsSourceMapTaskGradle53Plus>(name, configurationAction)
+                } else -> {
+                    project.tasks.register<BugsnagUploadJsSourceMapTaskLegacy>(name, configurationAction)
+                }
+            }
+        }
     }
+}
+
+/**
+ * Legacy [BugsnagUploadJsSourceMapTask] task that requires using [ProjectLayout.configurableFiles].
+ */
+internal open class BugsnagUploadJsSourceMapTaskLegacy @Inject constructor(
+    objects: ObjectFactory,
+    projectLayout: ProjectLayout
+) : BugsnagUploadJsSourceMapTask(objects) {
+
+    @get:InputFiles
+    override val projectRootFileProvider: ConfigurableFileCollection = projectLayout.configurableFiles()
+}
+
+internal open class BugsnagUploadJsSourceMapTaskGradle53Plus @Inject constructor(
+    objects: ObjectFactory
+) : BugsnagUploadJsSourceMapTask(objects) {
+
+    @get:InputFiles
+    override val projectRootFileProvider: ConfigurableFileCollection = objects.fileCollection()
 }