+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Create separate task for generating JVM mapping file
   [#335](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/335)
 
+* Compress JVM mapping files using gzip
+  [#336](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/336)
+
 ## 5.3.0 (2020-10-15)
 
 * Ensure libil2cpp has correct file extension